Refactor DateRange and DiscountType classes.
Move them into api module so that no circular dependencies
between modules exist.
