+++ api/pom.xml
@@ -13,8 +13,25 @@
     
     <dependencies>
         <dependency>
+            <groupId>org.assertj</groupId>
+            <artifactId>assertj-core</artifactId>
+            <version>3.11.1</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.testng</groupId>
+            <artifactId>testng</artifactId>
+            <scope>test</scope>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>cz.muni.fi.pa165</groupId>
-            <artifactId>service</artifactId>
+            <artifactId>persistence</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
     </dependencies>
