+++ api/src/main/java/cz/muni/fi/pa165/api/DateRange.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.pa165.service;
+package cz.muni.fi.pa165.api;
 
 import java.time.LocalDate;
