+++ api/src/main/java/cz/muni/fi/pa165/api/DiscountType.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.pa165.service;
+package cz.muni.fi.pa165.api;
 
 /**
  * Enumeration representing all possible discounts in the system.
