+++ service/src/main/java/cz/muni/fi/pa165/service/AdminServiceImpl.java
@@ -1,11 +1,12 @@
 package cz.muni.fi.pa165.service;
 
+import cz.muni.fi.pa165.api.DateRange;
 import cz.muni.fi.pa165.dao.BookingDao;
 import cz.muni.fi.pa165.entity.Booking;
 import cz.muni.fi.pa165.entity.Room;
 import cz.muni.fi.pa165.entity.User;
 import cz.muni.fi.pa165.service.exceptions.BookingManagerDataAccessException;
-import static cz.muni.fi.pa165.utils.DateRangeUtils.isBookingInsideDateRange;
+import static cz.muni.fi.pa165.api.utils.DateRangeUtils.isBookingInsideDateRange;
 import java.util.List;
 import java.util.stream.Collectors;
