+++ service/src/test/java/cz/muni/fi/pa165/service/DateRangeTest.java
@@ -1,5 +1,6 @@
 package cz.muni.fi.pa165.service;
 
+import cz.muni.fi.pa165.api.DateRange;
 import java.time.LocalDate;
 import java.time.Month;
 import org.junit.Test;
